# movie-web-app/server/routers/movieApp-API-docs.json
{
    "openapi": "3.0.3",
    "info": {
      "title": "MovieApp API documentation",
      "version": "1.0"
    },
    "servers": [
      {
        "url": "http://localhost:3001"
      }
    ],
    "paths": {
...
      "/groups": {
        "get": {
          "summary": "Get all groups",
          "description": "Retrieves all groups available in the system.",
          "responses": {
            "200": {
              "description": "A list of groups.",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "group_id": {
                          "type": "integer"
                        },
                        "name": {
                          "type": "string"
                        },
                        "description": {
                          "type": "string"
                        },
                        "owner_id": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error."
            }
          }
        },
        "post": {
          "summary": "Create a new group",
          "description": "Creates a new group with the provided name and description.",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "example": "My Movie Group"
                    },
                    "description": {
                      "type": "string",
                      "example": "A group for movie enthusiasts."
                    },
                    "owner_id": {
                      "type": "integer"
                    }
                  },
                  "required": ["name", "description", "owner_id"]
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Group successfully created.",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "group_id": {
                        "type": "integer"
                      },
                      "name": {
                        "type": "string"
                      },
                      "description": {
                        "type": "string"
                      },
                      "owner_id": {
                        "type": "integer"
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Validation error or duplicate group name."
            }
          }
        }
      },
      "/groups/{groupId}": {
        "get": {
          "summary": "Get group details",
          "description": "Retrieves details of a specific group by its ID.",
          "parameters": [
            {
              "name": "groupId",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Group details retrieved successfully.",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "group_id": {
                        "type": "integer"
                      },
                      "name": {
                        "type": "string"
                      },
                      "description": {
                        "type": "string"
                      },
                      "owner_id": {
                        "type": "integer"
                      }
                    }
                  }
                }
              }
            },
            "404": {
              "description": "Group not found."
            }
          }
        },
        "delete": {
          "summary": "Delete a group",
          "description": "Deletes a specific group by its ID. Only the group owner can delete it.",
          "parameters": [
            {
              "name": "groupId",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Group deleted successfully."
            },
            "403": {
              "description": "User is not authorized to delete the group."
            },
            "404": {
              "description": "Group not found."
            }
          }
        }
      },
...
    }
  }